requirements.txt
alembic==1.18.1
asyncpg==0.31.0
fastapi==0.128.0
fastapi-filter==2.0.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
SQLAlchemy==2.0.45
uvicorn==0.40.0
